+++ .NET/Microsoft.Recognizers.Text.Number/Parsers/BaseNumberParser.cs
@@ -181,6 +181,80 @@ namespace Microsoft.Recognizers.Text.Number
             return string.Join("|", sortKeys);
         }
 
+        protected static string DetermineType(ExtractResult er)
+        {
+            if (!string.IsNullOrEmpty(er.Type) && er.Type.Contains(Constants.MODEL_ORDINAL))
+            {
+                return er.Metadata.IsOrdinalRelative ? Constants.MODEL_ORDINAL_RELATIVE : Constants.MODEL_ORDINAL;
+            }
+
+            var data = er.Data as string;
+            var subType = string.Empty;
+
+            if (!string.IsNullOrEmpty(data))
+            {
+                if (data.StartsWith(Constants.FRACTION_PREFIX, StringComparison.Ordinal))
+                {
+                    subType = Constants.FRACTION;
+                }
+                else if (data.Contains(Constants.POWER_SUFFIX))
+                {
+                    subType = Constants.POWER;
+                }
+                else if (data.StartsWith(Constants.INTEGER_PREFIX, StringComparison.Ordinal))
+                {
+                    subType = Constants.INTEGER;
+                }
+                else if (data.StartsWith(Constants.DOUBLE_PREFIX, StringComparison.Ordinal))
+                {
+                    subType = Constants.DECIMAL;
+                }
+            }
+
+            return subType;
+        }
+
+        protected static bool IsMergeable(double former, double later)
+        {
+            // The former number is an order of magnitude larger than the later number, and they must be integers
+            return Math.Abs(former % 1) < double.Epsilon && Math.Abs(later % 1) < double.Epsilon &&
+                   former > later && former.ToString(CultureInfo.InvariantCulture).Length > later.ToString(CultureInfo.InvariantCulture).Length && later > 0;
+        }
+
+        // Test if big and combine with small.
+        // e.g. "hundred" can combine with "thirty" but "twenty" can't combine with "thirty".
+        protected static bool IsComposable(long big, long small)
+        {
+            var baseNumber = small > 10 ? 100 : 10;
+
+            return big % baseNumber == 0 && big / baseNumber >= 1;
+        }
+
+        protected double GetPointValue(List<string> matchStrs)
+        {
+            double ret = 0;
+            var firstMatch = matchStrs.First();
+
+            if (Config.CardinalNumberMap.ContainsKey(firstMatch) && Config.CardinalNumberMap[firstMatch] >= 10)
+            {
+                var prefix = "0.";
+                var tempInt = GetIntValue(matchStrs);
+                var all = prefix + tempInt;
+                ret = double.Parse(all, CultureInfo.InvariantCulture);
+            }
+            else
+            {
+                var scale = 0.1;
+                foreach (string matchStr in matchStrs)
+                {
+                    ret += Config.CardinalNumberMap[matchStr] * scale;
+                    scale *= 0.1;
+                }
+            }
+
+            return ret;
+        }
+
         protected ParseResult PowerNumberParse(ExtractResult extResult)
         {
             var result = new ParseResult
@@ -625,56 +699,7 @@ namespace Microsoft.Recognizers.Text.Number
             return calResult;
         }
 
-        private static string DetermineType(ExtractResult er)
-        {
-            if (!string.IsNullOrEmpty(er.Type) && er.Type.Contains(Constants.MODEL_ORDINAL))
-            {
-                return er.Metadata.IsOrdinalRelative ? Constants.MODEL_ORDINAL_RELATIVE : Constants.MODEL_ORDINAL;
-            }
-
-            var data = er.Data as string;
-            var subType = string.Empty;
-
-            if (!string.IsNullOrEmpty(data))
-            {
-                if (data.StartsWith(Constants.FRACTION_PREFIX, StringComparison.Ordinal))
-                {
-                    subType = Constants.FRACTION;
-                }
-                else if (data.Contains(Constants.POWER_SUFFIX))
-                {
-                    subType = Constants.POWER;
-                }
-                else if (data.StartsWith(Constants.INTEGER_PREFIX, StringComparison.Ordinal))
-                {
-                    subType = Constants.INTEGER;
-                }
-                else if (data.StartsWith(Constants.DOUBLE_PREFIX, StringComparison.Ordinal))
-                {
-                    subType = Constants.DECIMAL;
-                }
-            }
-
-            return subType;
-        }
-
-        private static bool IsMergeable(double former, double later)
-        {
-            // The former number is an order of magnitude larger than the later number, and they must be integers
-            return Math.Abs(former % 1) < double.Epsilon && Math.Abs(later % 1) < double.Epsilon &&
-                   former > later && former.ToString(CultureInfo.InvariantCulture).Length > later.ToString(CultureInfo.InvariantCulture).Length && later > 0;
-        }
-
-        // Test if big and combine with small.
-        // e.g. "hundred" can combine with "thirty" but "twenty" can't combine with "thirty".
-        private static bool IsComposable(long big, long small)
-        {
-            var baseNumber = small > 10 ? 100 : 10;
-
-            return big % baseNumber == 0 && big / baseNumber >= 1;
-        }
-
-        private string GetResolutionStr(object value)
+        protected string GetResolutionStr(object value)
         {
             var resolutionStr = value.ToString();
 
@@ -689,7 +714,7 @@ namespace Microsoft.Recognizers.Text.Number
         // Special cases for multi-language countries where decimal separators can be used interchangeably. Mostly informally.
         // Ex: South Africa, Namibia; Puerto Rico in ES; or in Canada for EN and FR.
         // "me pidio $5.00 prestados" and "me pidio $5,00 prestados" -> currency $5
-        private bool SkipNonDecimalSeparator(char ch, int distance)
+        protected bool SkipNonDecimalSeparator(char ch, int distance)
         {
             const int decimalLength = 3;
 
@@ -846,31 +871,6 @@ namespace Microsoft.Recognizers.Text.Number
             }
 
             return tempValue;
-        }
-
-        private double GetPointValue(List<string> matchStrs)
-        {
-            double ret = 0;
-            var firstMatch = matchStrs.First();
-
-            if (Config.CardinalNumberMap.ContainsKey(firstMatch) && Config.CardinalNumberMap[firstMatch] >= 10)
-            {
-                var prefix = "0.";
-                var tempInt = GetIntValue(matchStrs);
-                var all = prefix + tempInt;
-                ret = double.Parse(all, CultureInfo.InvariantCulture);
-            }
-            else
-            {
-                var scale = 0.1;
-                foreach (string matchStr in matchStrs)
-                {
-                    ret += Config.CardinalNumberMap[matchStr] * scale;
-                    scale *= 0.1;
-                }
-            }
-
-            return ret;
         }
 
         private Regex BuildTextNumberRegex()
